Main: fix win burns
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@13850 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/testcase/tstAPI.cpp
@@ -776,7 +776,7 @@ int main(int argc, char *argv[])
             CHECK_ERROR_BREAK (hardDiskFormats [i],
                                COMGETTER(Id) (description.asOutParam()));
 
-            unsigned int caps;
+            ULONG caps;
             CHECK_ERROR_BREAK (hardDiskFormats [i],
                                COMGETTER(Capabilities) (&caps));
